Refactored GoogleReviewSection
-Moved components and functions to separate files
-Added types for props
-Memoized the carousel to prevent rerenders
